requirements.txt
# Core dependencies for the perceptron image classification project
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.3.0

